refactor: getters 파일 분리
